+++ Contents/Code/PAsearchSites.py
@@ -79,7 +79,7 @@ import siteStepSecrets
 import siteVRHush
 import networkMetArt
 
-searchSites = [None] * 755
+searchSites = [None] * 757
 
 searchSites[1] = ["Blacked com","Blacked","https://www.blacked.com","https://www.blacked.com/search?q="]
 searchSites[0] = ["Blackedraw com","BlackedRaw","https://www.blackedraw.com","https://www.blackedraw.com/search?q="]
@@ -836,6 +836,8 @@ searchSites[751] = ["Step Secrets","Step Secrets","http://www.stepsecrets.com","
 searchSites[752] = ["VRHush","VRHush","https://www.vrhush.com","https://vrhush.com/scenes/"]
 searchSites[753] = ["MetArt","MetArt","https://www.metart.com","https://www.metart.com/search/"]
 searchSites[754] = ["MetArtX","MetArtX","https://www.metartx.com","https://www.metartx.com/search/"]
+searchSites[755] = ["NubilesET","NubilesET","https://nubileset.com/","https://nubileset.com/video/"]
+searchSites[756] = ["Detention Girls","Detention Girls","https://detentiongirls.com","https://detentiongirls.com/video/"]
 
 
 def getSearchBaseURL(siteID):